added equivalence checks and separated xml2d
